# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dropolis API - Portal for Dropolë villages
 * OpenAPI spec version: 0.1.0
 */
export interface ChatPresence {
  online: number;
}

export interface ChatPresenceInput {
  username: string;
}

export interface HealthStatus {
  status: string;
}

export interface Article {
  id: number;
  title: string;
  /** @nullable */
  excerpt?: string | null;
  content: string;
  category: string;
  author: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  villageName?: string | null;
  /** @nullable */
  tags?: string | null;
  /** @nullable */
  seoTitle?: string | null;
  /** @nullable */
  metaDescription?: string | null;
  /** @nullable */
  slug?: string | null;
  score?: number;
  viewCount?: number;
  published: boolean;
  featured?: boolean;
  createdAt: string;
  updatedAt?: string;
}

export interface ArticleInput {
  title: string;
  excerpt?: string;
  content: string;
  category: string;
  author: string;
  imageUrl?: string;
  villageName?: string;
  tags?: string;
  seoTitle?: string;
  metaDescription?: string;
  slug?: string;
  published?: boolean;
  featured?: boolean;
}

export interface ArticleUpdate {
  title?: string;
  excerpt?: string;
  content?: string;
  category?: string;
  author?: string;
  imageUrl?: string;
  villageName?: string;
  tags?: string;
  seoTitle?: string;
  metaDescription?: string;
  slug?: string;
  score?: number;
  published?: boolean;
  featured?: boolean;
}

export interface SocialPost {
  articleId: number;
  articleTitle: string;
  fbPost: string;
  shortCaption: string;
  hashtags: string[];
}

export interface Category {
  name: string;
  count: number;
}

export interface Village {
  id: number;
  name: string;
  nameEl: string;
  description: string;
  /** @nullable */
  municipalUnit?: string | null;
  /** @nullable */
  population: number | null;
  /** @nullable */
  elevation?: number | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  createdAt?: string;
}

export interface VillageInput {
  name: string;
  nameEl: string;
  description: string;
  municipalUnit?: string;
  population?: number;
  elevation?: number;
  imageUrl?: string;
  latitude?: number;
  longitude?: number;
}

export interface VillageUpdateInput {
  name?: string;
  nameEl?: string;
  description?: string;
  municipalUnit?: string;
  population?: number;
  elevation?: number;
  imageUrl?: string | null;
  latitude?: number;
  longitude?: number;
}

export interface Photo {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  url: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  villageId?: number | null;
  /** @nullable */
  villageName?: string | null;
  /** @nullable */
  photographer?: string | null;
  createdAt: string;
}

export interface PhotoInput {
  title: string;
  description?: string;
  url: string;
  thumbnailUrl?: string;
  villageId?: number;
  photographer?: string;
}

export interface PhotoUploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export interface PhotoSubmission {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  objectPath: string;
  villageId?: number;
  villageName?: string;
  photographer?: string;
  uploaderName?: string;
  copyrightConfirmed: boolean;
}

export interface AdminPhoto {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  url: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  villageId?: number | null;
  /** @nullable */
  villageName?: string | null;
  /** @nullable */
  photographer?: string | null;
  status: string;
  /** @nullable */
  objectPath?: string | null;
  copyrightConfirmed: boolean;
  /** @nullable */
  uploaderName?: string | null;
  createdAt: string;
}

export interface ErrorEnvelope {
  error: string;
}

export interface Video {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  youtubeId: string;
  /** @nullable */
  villageId?: number | null;
  /** @nullable */
  villageName?: string | null;
  /** @nullable */
  duration?: string | null;
  createdAt: string;
}

export interface VideoInput {
  title: string;
  description?: string;
  youtubeId: string;
  villageId?: number;
  duration?: string;
}

export interface ChatMessage {
  id: number;
  username: string;
  message: string;
  /** @nullable */
  avatar?: string | null;
  isBot?: boolean;
  createdAt: string;
}

export interface ChatMessageInput {
  username: string;
  message: string;
  avatar?: string;
}

export interface SiteStats {
  totalArticles: number;
  totalVillages: number;
  totalPhotos: number;
  totalVideos: number;
  totalMessages: number;
  recentArticles?: number;
  featuredCount?: number;
}

export type ListArticlesParams = {
category?: string;
village?: string;
limit?: number;
offset?: number;
};

export type GetTrendingArticlesParams = {
limit?: number;
};

export type RecalculateArticleScores200 = {
  ok: boolean;
  updated: number;
};

export type ListPhotosParams = {
village_id?: number;
limit?: number;
offset?: number;
};

export type ListAdminPhotosParams = {
status?: string;
};

export type ListVideosParams = {
village_id?: number;
limit?: number;
};

export type ListChatMessagesParams = {
limit?: number;
};

export type DeleteChatMessageParams = {
username: string;
};