# lib/api-zod/src/generated/types/adminPhoto.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dropolis API - Portal for Dropolë villages
 * OpenAPI spec version: 0.1.0
 */

export interface AdminPhoto {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  url: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  villageId?: number | null;
  /** @nullable */
  villageName?: string | null;
  /** @nullable */
  photographer?: string | null;
  status: string;
  /** @nullable */
  objectPath?: string | null;
  copyrightConfirmed: boolean;
  /** @nullable */
  uploaderName?: string | null;
  createdAt: string;
}