# lib/api-zod/src/generated/types/photoSubmission.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dropolis API - Portal for Dropolë villages
 * OpenAPI spec version: 0.1.0
 */

export interface PhotoSubmission {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  objectPath: string;
  villageId?: number;
  villageName?: string;
  photographer?: string;
  uploaderName?: string;
  copyrightConfirmed: boolean;
}