# lib/api-zod/src/generated/types/photoUploadUrlRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dropolis API - Portal for Dropolë villages
 * OpenAPI spec version: 0.1.0
 */

export interface PhotoUploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}