# lib/api-zod/src/generated/types/uploadUrlResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dropolis API - Portal for Dropolë villages
 * OpenAPI spec version: 0.1.0
 */

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}